+++ bundles/framework/bundle/mapfull/locale/fi.js
@@ -1,0 +1,7 @@
+Oskari.registerLocalization({
+    "lang" : "fi",
+    "key" : "Generic",
+    "value" : {
+        "default-style": "Oletustyyli"
+    }
+});